Move main and parse_arge to a stand-alone script
